feat: functions and main.py on root
